adds users invitations and registration function (#6)
* reactors project structure - moves handlers to separate directories
* (wip) adds user registration (empty register cloud function)
* (wip) adds user registration (use FieldPath for firestore query)
* (wip) adds user registration (removes temporary test code)
* (wip) adds user registration (adds user creation for valid invitations)
* (wip) adds user registration (changes system user creation to not delete unknown users [to protect manually created users])

+++ tools/users-sync/users-sync.ts
@@ -34,15 +34,19 @@ const chapterLeaderMap = users.users
 (async function () {
   console.log(`Update users with chapter link`);
   await updateCollection(`team/${team}/user/`, userMap, user => ({
-    chapterLeader: userMap[user].chapterLeader
-  }));
+      chapterLeader: userMap[user].chapterLeader
+    }),
+    true
+  );
 
   console.log(`Update system users`);
   await updateCollection(`user/`, chapterLeaderMap, user => ({
-    teams: {
-      [team]: true
-    }
-  }));
+      teams: {
+        [team]: true
+      }
+    }),
+    false
+  );
 })();
 
 
@@ -50,7 +54,7 @@ function onlyUnique(value, index, self) {
   return self.indexOf(value) === index;
 }
 
-async function updateCollection(collection, users, updateFn: (user) => any) {
+async function updateCollection(collection, users, updateFn: (user) => any, removeUnknown: boolean) {
   const userCollection = firebase.firestore().collection(collection);
   const existingUsers = await userCollection.get();
   const usersToSync = Object.keys(users).length;
@@ -61,7 +65,7 @@ async function updateCollection(collection, users, updateFn: (user) => any) {
         userCollection.doc(existingUser.id).update(updateFn(existingUser.id));
       }
       delete users[existingUser.id];
-    } else {
+    } else if (removeUnknown) {
       console.log(`  remove user ${existingUser.id}`);
       userCollection.doc(existingUser.id).delete();
     }
